full_requirements.txt
Flask==0.12.1
picamera==1.13
pyTelegramBotAPI==3.6.6
RPi.GPIO==0.6.5
ws4py==0.5.1
PyYAML==5.1
SQLAlchemy==1.3.5
celery==4.3.0
requests==2.21.0
numpy==1.16.4
opencv-python==3.4.4.19
tensorflow==1.14.0